+++ yantra-canvas/dribbble_com/pages/index.tsx
@@ -1,7 +1,0 @@
-
-export default function IndexPage() {
-  return (
-    <div data-yantra-id="230d8358dc8e">
-    </div>
-  );
-}